Rename satpy.config to satpy._config
