+++ satpy/_config.py
