+++ satpy/readers/eps_l1b.py
@@ -27,7 +27,7 @@ import xarray as xr
 import dask.array as da
 from dask.delayed import delayed
 from pyresample.geometry import SwathDefinition
-from satpy.config import CONFIG_PATH
+from satpy._config import CONFIG_PATH
 from satpy.readers.file_handlers import BaseFileHandler
 from satpy.readers.xmlformat import XMLFormat
 from satpy import CHUNK_SIZE
