+++ satpy/tests/reader_tests/test_agri_l1.py
@@ -73,7 +73,7 @@ class FakeHDF5FileHandler2(FakeHDF5FileHandler):
                                     'Slope': 1., 'Intercept': 0.,
                                     'FillValue': 0,
                                     'units': 'NUL',
-                                    'band_names': 'NUL'.format(ch).encode('utf-8'),
+                                    'band_names': 'NUL',
                                     'long_name': b'Calibration coefficient (SCALE and OFFSET)',
                                     'valid_range': [-500, 500],
                                 },
@@ -86,7 +86,7 @@ class FakeHDF5FileHandler2(FakeHDF5FileHandler):
                                     'Slope': 1., 'Intercept': 0.,
                                     'FillValue': 0,
                                     'units': 'NUL',
-                                    'band_names': 'NUL'.format(ch).encode('utf-8'),
+                                    'band_names': 'NUL',
                                     'long_name': b'Calibration coefficient (SCALE and OFFSET)',
                                     'valid_range': [-500, 500],
                                 },
@@ -99,7 +99,7 @@ class FakeHDF5FileHandler2(FakeHDF5FileHandler):
                                     'Slope': 1., 'Intercept': 0.,
                                     'FillValue': 0,
                                     'units': 'NUL',
-                                    'band_names': 'NUL'.format(ch).encode('utf-8'),
+                                    'band_names': 'NUL',
                                     'long_name': b'Calibration coefficient (SCALE and OFFSET)',
                                     'valid_range': [-500, 500],
                                 },
@@ -113,7 +113,7 @@ class FakeHDF5FileHandler2(FakeHDF5FileHandler):
                                     'Slope': 1., 'Intercept': 0.,
                                     'FillValue': 0,
                                     'units': 'NUL',
-                                    'band_names': 'NUL'.format(ch).encode('utf-8'),
+                                    'band_names': 'NUL',
                                     'long_name': b'Calibration coefficient (SCALE and OFFSET)',
                                     'valid_range': [-500, 500],
                                 },
@@ -228,7 +228,7 @@ class Test_HDF_AGRI_L1_cal(unittest.TestCase):
     def setUp(self):
         """Wrap HDF5 file handler with our own fake handler."""
         from satpy.readers.agri_l1 import HDF_AGRI_L1
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         # http://stackoverflow.com/questions/12219967/how-to-mock-a-base-class-with-python-mock-library
         self.p = mock.patch.object(HDF_AGRI_L1, '__bases__', (FakeHDF5FileHandler2,))
