+++ satpy/tests/reader_tests/test_amsr2_l2_gaasp.py
@@ -171,7 +171,7 @@ class TestGAASPReader:
 
     def setup_method(self):
         """Wrap pygrib to read fake data."""
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
 
     @pytest.mark.parametrize(
