+++ satpy/tests/reader_tests/test_clavrx.py
@@ -112,7 +112,7 @@ class TestCLAVRXReaderPolar(unittest.TestCase):
 
     def setUp(self):
         """Wrap HDF4 file handler with our own fake handler."""
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         from satpy.readers.clavrx import CLAVRXFileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         # http://stackoverflow.com/questions/12219967/how-to-mock-a-base-class-with-python-mock-library
@@ -304,7 +304,7 @@ class TestCLAVRXReaderGeo(unittest.TestCase):
 
     def setUp(self):
         """Wrap HDF4 file handler with our own fake handler."""
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         from satpy.readers.clavrx import CLAVRXFileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         # http://stackoverflow.com/questions/12219967/how-to-mock-a-base-class-with-python-mock-library
