+++ satpy/tests/reader_tests/test_glm_l2.py
@@ -209,7 +209,7 @@ class TestGLML2Reader(unittest.TestCase):
     def setUp(self, xr_):
         """Create a fake reader to test."""
         from satpy.readers import load_reader
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         fake_dataset = setup_fake_dataset()
         xr_.open_dataset.return_value = fake_dataset
