+++ satpy/tests/reader_tests/test_mimic_TPW2_nc.py
@@ -90,7 +90,7 @@ class TestMimicTPW2Reader(unittest.TestCase):
 
     def setUp(self):
         """Wrap NetCDF4 file handler with our own fake handler."""
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         from satpy.readers.mimic_TPW2_nc import MimicTPW2FileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         # http://stackoverflow.com/questions/12219967/how-to-mock-a-base-class-with-python-mock-library
