+++ satpy/tests/reader_tests/test_mviri_l1b_fiduceo_nc.py
@@ -337,7 +337,7 @@ def fixture_file_handler(fake_dataset, request):
 @pytest.fixture(name='reader')
 def fixture_reader():
     """Return MVIRI FIDUCEO FCDR reader."""
-    from satpy.config import config_search_paths
+    from satpy._config import config_search_paths
     from satpy.readers import load_reader
 
     reader_configs = config_search_paths(
