+++ satpy/tests/reader_tests/test_smos_l2_wind.py
@@ -80,7 +80,7 @@ class TestSMOSL2WINDReader(unittest.TestCase):
 
     def setUp(self):
         """Wrap NetCDF4 file handler with our own fake handler."""
-        from satpy.config import config_search_paths
+        from satpy._config import config_search_paths
         from satpy.readers.smos_l2_wind import SMOSL2WINDFileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         # http://stackoverflow.com/questions/12219967/how-to-mock-a-base-class-with-python-mock-library
